Break out funding cycle formatting utils to be reused
